+++ crates/kftray-kube/src/error.rs
