+++ crates/kftray-kube/src/kube/stop/all.rs
@@ -35,7 +35,7 @@ use super::cleanup::{
     load_configs,
 };
 use crate::shared_client::ServiceClientKey;
-use crate::port_forward_error::PortForwardError;
+use crate::error::PortForwardError;
 use crate::registry::PORT_FORWARD_REGISTRY;
 
 pub async fn stop_all_port_forward() -> Result<Vec<CustomResponse>, PortForwardError> {
